pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.9.0"
description = ""
authors = [
    {name = "Artur Didur",email = "didur.art@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
]

[tool.poetry.scripts]
start = "app.main:main"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"
aiosqlite = "^0.21.0"

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
